replace NotationParser.parse with Notation.parse;
delete NotationParser;
